django-drf/backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.1; python_version >= '3.10'
django-autoslug==1.9.9
django-cors-headers==4.4.0; python_version >= '3.8'
django-debug-toolbar==4.4.6; python_version >= '3.8'
django-filter==24.3; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-nested-routers==0.94.1; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyjwt==2.9.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
whitenoise==6.7.0; python_version >= '3.8'
